+++ entity/src/main/resources/db/migration/tables/user/V1.55__user_indexes.sql
@@ -1,0 +1,11 @@
+-- ------------------------------------------------
+-- Version: v1.55
+-- -------------------------------------------------
+
+CREATE INDEX idx_email ON User (email);
+CREATE INDEX idx_orcid_id ON User (orcid_id);
+CREATE INDEX idx_name ON User (name);
+CREATE INDEX idx_surname ON User (surname);
+
+
+